Avoid taking a lock in `PollEvented::poll_ready`
This function is the source of all readiness notifications (and polling) in
`PollEvented`. Currently, however, it'll bounce on an rwlock (in read mode),
which can often cause performance degredation across platforms.

This commit instead alters the logic to avoid hitting a lock in `poll_ready`
entirely. This is achieve by storing `Arc<ScheduledIo>`, the allocated state for
each I/O object, inside the I/O objects directly. Previously only an index
(`token: usize`) was stored which meant that the original slab had to be
reindexed (and locked) each time an I/O object needed to look up its state.

This does come at the cost of a forced allocation per I/O object, but that was
effectively already happening in the previous iteration with allocation space in
the slab. If this ends up becoming a problem then we can always look into
caching mechanisms as well for `Arc<ScheduledIo>` and have a slab of free
versions on the reactor.